Changed validation flow to start file meta validation without the need for cloudUrl to be present
